Document and clean-up certain battle-pause-related sections...
Auto-pause doesn't work in multiplayer, for-example.
I'm tempted to properly replace pause with battle-pause but that would require removing 40 references to it. The battle-pause code was able to automatically-merge with 600 newer up-stream commits because it is so short (and clever in side-stepping the original pause-function). As such: I'm leaving things as-is for maintanabilty's sake. If anything, the extra clean-up should make it even less likely to conflict with up-stream changes.